feat(internal): Refactor to provide base Agent class & allow for alternate runtimes (#146)
